Remove speed argument from baton_init
Prior to this commit, the baton_init() function took an argument for the
speed of the serial communication, but according to the Numato USB Relay
documentation, the USB controller on the relay board doesn't care what
speed it's set to as long as it's a valid baudrate. This commit removes
the speed argument to simplify the usage; it uses a standard baud rate
of 19200, which is what the examples in the documentation use.