Add sleep after read() and write() calls
Prior to this commit, the library would write bytes to the relay and
read bytes from the relay with no delay after read() and write(), which
would cause occasional hangups when the relay was being overloaded and
could not respond. This commit adds 1ms sleeps after calls to read() and
write() to give the relay a moment to respond.