Check for operation success after completing
Prior to this commit, operations that were expected to change the
relay's state (i.e., enabling of relay, disabling of relay, setting
module id, and toggling by bitfield) did not query the device afterward
to read back the expected state change. This commit modifies the
state-changing operations to query the module afterwrad to ensure that
the requested operation did what it was expected to do.